A few proofing updates.  General comments.
statue -> statute
federal government -> Federal Government
some whitespace edits

Nice work, Jack. I don't see much that needs to be added, but it might be expanded a bit some other policies that LBJ put into effect during his Great Society run. One that comes to mind is the Americorps VISTA program. We used a VISTA at KDHX to help with a massive, entirely internal rebuild of the web presence in 08-09. It was a yearlong project that may not have been completed if it weren't for LBJ's programs. While there wasn't a portion of the VISTA program focused on digital technology for public service when it was formed, the framework evolved with the needs, much as public media has.

Maybe it should now be called the Corporation for Public Media? Except, ouch, CPM is kinda taken, and loaded.